rng v0.4 for geometric distribution

+++ src/rng/bernoulli.go
@@ -29,9 +29,9 @@ func (beng BernoulliGenerator) Bernoulli() bool {
 }
 
 // Bernoulli_P returns a bool, which is true with probablity p
-func (beng BernoulliGenerator) Bernoulli_P(p float32) bool {
+func (beng BernoulliGenerator) Bernoulli_P(p float64) bool {
 	if !(0.0 <= p && p <= 1.0) {
 		panic(fmt.Sprintf("Invalid probability: %.2f", p))
 	}
-	return beng.uniform.Float32() < p
+	return beng.uniform.Float64() < p
 }
